#3186: consolidate svm, svr dataset validation
