--- brain/schema/__init__.py
@@ -1,4 +1,0 @@
-'''
-This file allows the containing directory to be considered a python package,
-consisting of python module(s).
-'''
